packages/edict/pyproject.toml
[project]
name = "edict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Massimo", email = "massimo@adimian.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.11.0",
]

[project.scripts]
edict = "edict:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
